src/test/javascript/spec/app/entities/fiche-emprunt-produit/fiche-emprunt-produit.service.spec.ts
/* tslint:disable max-line-length */
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { HttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { take, map } from 'rxjs/operators';
import * as moment from 'moment';
import { DATE_FORMAT } from 'app/shared/constants/input.constants';
import { FicheEmpruntProduitService } from 'app/entities/fiche-emprunt-produit/fiche-emprunt-produit.service';
import { IFicheEmpruntProduit, FicheEmpruntProduit } from 'app/shared/model/fiche-emprunt-produit.model';

describe('Service Tests', () => {
    describe('FicheEmpruntProduit Service', () => {
        let injector: TestBed;
        let service: FicheEmpruntProduitService;
        let httpMock: HttpTestingController;
        let elemDefault: IFicheEmpruntProduit;
        let currentDate: moment.Moment;
        beforeEach(() => {
            TestBed.configureTestingModule({
                imports: [HttpClientTestingModule]
            });
            injector = getTestBed();
            service = injector.get(FicheEmpruntProduitService);
            httpMock = injector.get(HttpTestingController);
            currentDate = moment();

            elemDefault = new FicheEmpruntProduit(0, 0, currentDate);
        });

        describe('Service methods', async () => {
            it('should find an element', async () => {
                const returnedFromService = Object.assign(
                    {
                        dateEmprunt: currentDate.format(DATE_FORMAT)
                    },
                    elemDefault
                );
                service
                    .find(123)
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: elemDefault }));

                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should create a FicheEmpruntProduit', async () => {
                const returnedFromService = Object.assign(
                    {
                        id: 0,
                        dateEmprunt: currentDate.format(DATE_FORMAT)
                    },
                    elemDefault
                );
                const expected = Object.assign(
                    {
                        dateEmprunt: currentDate
                    },
                    returnedFromService
                );
                service
                    .create(new FicheEmpruntProduit(null))
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'POST' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should update a FicheEmpruntProduit', async () => {
                const returnedFromService = Object.assign(
                    {
                        quantite: 1,
                        dateEmprunt: currentDate.format(DATE_FORMAT)
                    },
                    elemDefault
                );

                const expected = Object.assign(
                    {
                        dateEmprunt: currentDate
                    },
                    returnedFromService
                );
                service
                    .update(expected)
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'PUT' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should return a list of FicheEmpruntProduit', async () => {
                const returnedFromService = Object.assign(
                    {
                        quantite: 1,
                        dateEmprunt: currentDate.format(DATE_FORMAT)
                    },
                    elemDefault
                );
                const expected = Object.assign(
                    {
                        dateEmprunt: currentDate
                    },
                    returnedFromService
                );
                service
                    .query(expected)
                    .pipe(
                        take(1),
                        map(resp => resp.body)
                    )
                    .subscribe(body => expect(body).toContainEqual(expected));
                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify([returnedFromService]));
                httpMock.verify();
            });

            it('should delete a FicheEmpruntProduit', async () => {
                const rxPromise = service.delete(123).subscribe(resp => expect(resp.ok));

                const req = httpMock.expectOne({ method: 'DELETE' });
                req.flush({ status: 200 });
            });
        });

        afterEach(() => {
            httpMock.verify();
        });
    });
});
